S05P12C103-79 feat: [BE] 채팅 백엔드 구현 / erd테이블 맞춰서 수정

+++ backend/src/main/java/com/ssafy/db/repository/Live/LiveRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  * 방송 모델 관련 디비 쿼리 생성을 위한 JPA Query Method 인터페이스 정의.
  */
 @Repository
-public interface LiveRepository extends JpaRepository<Live, Long> {
+public interface LiveRepository extends JpaRepository<Live, Integer> {
     // 아래와 같이, Query Method 인터페이스(반환값, 메소드명, 인자) 정의를 하면 자동으로 Query Method 구현됨.
 //    Optional<Live> findByUserId(String userId);
 }
